Fix issue with "Unhinged" cards that have fractional CMC
Fix issue with "Unhinged" cards that have fractional CMC

+++ NerdBot/NerdBot/Mtg/Card.cs
@@ -50,7 +50,7 @@ namespace NerdBot.Mtg
 
         [JsonProperty("cmc")]
         [BsonElement("cmc")]
-        public int Cmc { get; set; }
+        public double Cmc { get; set; }
 
         [JsonProperty("setName")]
         [BsonElement("setName")]
